--- src/task/mod.rs
@@ -1,5 +1,0 @@
-mod base;
-mod spec;
-
-pub use base::{Task, TaskFn, TaskRef};
-pub use spec::TaskSpec;
