+++ src/tasks/mod.rs
@@ -1,0 +1,7 @@
+mod spec;
+mod task;
+mod task_fn;
+
+pub use spec::TaskSpec;
+pub use task::Task;
+pub use task_fn::{TaskFn, TaskRef};
